Touch up login message
Add some emotion-based text to the login messages, thanking volunteers
for their service, and hopefully providing some encouragement and a
human touch.

The text for the text message and the email is a bit different.

The text message text is much shorter, trying not to take up too much
space in a narrow chat bubble. It also has "Astoria Mutual Aid" first,
to better guarantee that it will appear in message summaries, so folks
will hopefully have a better sense of the nature of the text at a
glance.

The email version is a bit wordier, but not by much. Since we have the
luxury of clearly stating the sender in the subject line, and email
having more room to breath than text in general, use more natural
language to thank the volunteers.